# [Bugfix][VIP] Add workaround for FA2 inf with mini overhead